+++ release-notes.md
@@ -9,6 +9,7 @@ for [Chrome extension to for Bible.com](README.md).
 - [x] Duplicate Slide with Context Menu (right click)
 - [x] Check duplicate content when import settings
 - [x] Improve tables design in projector view (live text with markdown - add borders, padding)
+- [x] Allow more spaces to be displayed in live text markdown
  
 ## 2.3.0