+++ release-notes.md
@@ -4,7 +4,11 @@ for [Chrome extension to for Bible.com](README.md).
 
 💚 Love this extension? Share [feedback](https://chromewebstore.google.com/detail/project-verses-from-bible/fklnkmnlobkpoiifnbnemdpamheoanpj) and help us make it even better!
 
-## 2.4.0 (🔜 coming soon)
+## 2.5.0 (🔜 coming soon)
+
+- [ ] ...?
+
+## 2.4.0
 
 - [x] Duplicate Slide with Context Menu (right click)
 - [x] Check duplicate content when import settings